Revert "export hook" behaviour to the way it was before bb9eb667e2.
In particular, the client code runs before any kernel code is run. My
change thoughtlessly had the client code run after the theory was
"scrubbed" to remove it of out-of-date theorems. But Magnus's use of
the export hook in ml_translatorLib called delete_const resulting in a
bad theory export.

The boss/theory_tests change mimicks that problem in miniature.

This is fallout from issue #73.