update logging-build-sequence
There have been various changes to the build sequence not yet
incorporated, the most important being src/prob being renamed. In
general logging-build-sequence should just be a copy of build-sequence
with the KERNEL line replaced with src/logging-kernel, and maybe that
should be dealt with programatically (option to build, or script to
generate the sequence) some day.