Revise more of Parse's pretty-printing APIs.
In particular, remove a bunch of entry-points with names matching
*backend_string* and print_backend*.

The objective with these was to accompany things like
term_to_string (which printed to a raw_terminal backend) with a
printer that printed to the current backend. This is now easy enough
to do with

  ppstring pp_term

and

  print o ppstring pp_term

The flip of the current_backend reference so that it points at
"raw_terminal" can now also be done with the rawterm_pp function.

The other convenience entry-points that have gone, for the moment, are
those that did all of the above while also removing certain specified
overloads. Because pp_term_without_overloads_on is still present, you
have to make do with

   print (ppstring (pp_term_without_overloads_on ["/\\"]) ``p /\ q``);

or

   rawterm_pp (ppstring (pp_term_without_overloads_on ["/\\"]))
              ``p /\ q``;

Closes #13. I will open a fresh issue to do with making the last two
use-cases slightly easier to deal with.