rewrite labelise explicitly rather than using FOLDL
Termination and other (esp. inductive) proofs are simpler if one works
with TFL rather than against it... I keep running into this. In the end
I probably want the FOLDL because the compiler might be able to do
something clever, but it will still be easier to prove something
complicated about the expanded version then prove that function equal to
the FOLDL version.