start on label_closures semantics preservation thm
define free_bods/free_labs to take the code_env as a parameter (and
implicitly flatten things out, like free_vars does)

also define the flattening out function: a useful stepping stone will be
to say that Cevaluate is equivalent to Cevaluate FEMPTY of the flattened
expression.

attempt to prove that result, but it looks like it's going to be really
awful and require syneq etc. it may be worth going back and rethinking
the intLang proofs (and definitions) first.