Add ordinal theorems for exponentiation.
In particular:
  |- x ** (y * z) = (x ** z) ** y
  |- x ** (y + z) = x ** z * x ** y

The reversal of the y and z stems from the way multiplication is
defined by recursion on the first argument. If you defined
multiplication by recursion in the second argument, you have that

  2 * ω = ω

and

  ω * 2 = ω + ω

(rather than the reverse).

I prefer 2 * ω = ω + ω when imagining the writing of polynomials,
where

  ω ** 2 + 2 * ω + 1

seems more familiar.