move labelling of closure bodies into exp_to_Cexp
i.e. make the intermediate language always with respect to a code
environment, and no possibility of inlined bodies.

in this commit: got the compiler tests working with the new setup.
commented out one test that wrote an intermediate language expression
directly: this is no longer possible when there are any function bodies
in the expression.

commented out the definition of v_to_Cv, since it will now be much more
involved, and it may be better to use a relation between vs and Cvs
(like syneq) instead.

exp_to_Cexp proof is of course broken with v_to_Cv gone.