Make win-config's handling of sigobj/SRCFILES idempotent.
The make_iss script (before installation) turns the absolute paths in this file into relative ones, and
then win-config, which runs as part of installation, makes them absolute paths again.  But it's useful
to be able to run win-config repeatedly, so I've changed it so that if it sees absolute paths, it just leaves
them alone and doesn't try to change them (assuming that it's already done the right thing once).