Tightening humble type definitions, revised package axioms, better rank inf.
Some new conditions are added to tighten eligibility for a new type
constant being defined as a humble type constant; in particular, the
subset predicate must not involve any universal or existential types.

The package axioms are revised; in particular, the axiom PACK_EXT_AX
is removed, as being not in accord with the current version of the
set theory semantics. PACK_EXT_AX would not have introduced inconsistency,
but it would have needed a more complex semantics to justify it.

Rank inference is strengthened a bit to avoid losing information when
comparing unification variables to be <= successor rank expressions.