Automatically prove a bunch of “applied” results for set constants.
An “applied” theorem is of the form

  (set construction) x <=> ...

where the ... will be from the theorem

  x ∈ (set construction) <=> ...

So, for example, INSERT_applied is

  |- (x INSERT s) y <=> (x = y) \/ y ∈ s

These theorems are all exported as rewrites, and provide a way of
getting users out of holes where they have “set world” constructions
but have lost the ∈ connection between the set and the possible
element. Note that these theorems will attempt to put the users back
into the “set world” because the RHSs will retain their uses of ∈.

These theorems will still work fine when rewriting with IN_DEF:
something like

   SIMP_TAC (srw_ss()) [IN_DEF]

will eliminate the “IN”s, even as the “applied” theorems help to make
progress with the other bits of the goal.

There’s no handling of GSPEC yet, but perhaps that can wait for HOL
Light style set-comprehension (see issue #85). The alternative is
probably a special conversion to do the right thing, though perhaps
just rewriting with

  GSPEC f v <=> ∃x. (v,T) = f x

would do the trick in most cases. (The problem in general is that to
get the above rewrite to work you have to throw in
pairTheory.EXISTS_PROD as well.)