Hacky fix for Holmake build errors in src/emit/ML.
I had forgotten that the ordinal example causes src/emit/ML to be
looked at even when the selftest level is 1.  I had assumed that
Holmake was only going to be called in this directory when part
of a selftest 2 build.  The longer-term solution will be to
set Holmake up so that it can receive variable bindings on the
command-line (as regular make can), and then arrange to invoke it
that way by build, with for example

  Holmake SELFTESTLEVEL=1