Exit with non-zero code if configuration fails under Poly/ML.
Issue is that you can get a Fail exception when a call to ‘use’ fails,
but this doesn’t translate into a non-zero exit code when the Poly/ML
process exits, and this means that the regression build thinks that
configuration has succeeded, and then the build keeps going. Which is
bad. This (and the fact that Poly/ML 5.20 can’t compile
tools-poly/configure.sml) explains the regressions (with commit
7c7739a8c and earlier).