Remove LIST_TO_SET from emit’s listML.
Having LIST_TO_SET present as a function of type α list → (α → bool)
causes a clash with expected use of that function targetting a
set-type. Perhaps the function should be of type α list → α set (where
the range of this is not just α → bool in the emitted ML world).
However, as this function was not present previously, I guess there is
no demand for it, and think it’s probably just easiest to leave it
out.

Closes #52. See also github comments on f0ab91f for some more
discussion.