Start of a simple theory about parsing and context free languages.
Specify “line-by-line” derivations with respect to a grammar (a
derivation takes the start symbol as the first line, and on each
successive line expands one non-terminal), as well as a parse-tree
approach.

Show that both notions coincide in their notion of a grammar’s language.

This much is inspired by Aditi Barthwal’s PhD thesis.

Next step is to look at PEGs, as a possible basis for more efficient,
easily executable parsing. PEGs should produce parse trees that are of
the same type as the parse trees in this theory.