Remove tfl/src/Functional.{sig,sml}; redirect to src/1/Pmatch instead.
This preliminary check-in doesn’t bring across all of the differences
between Functional and Pmatch. On the assumption that the differences
are important, I’ll patch Pmatch as necessary on this branch before
merging back into master.

Intention is to have this done before proceeding with work on
features/case-const-argflip. Given that, this is progress of a sort
with Github issue #97.