Make sum_case_tm have its type variables allocated as before.
If you write ``sum_CASE``, you get a term with type variables
allocated (alpha, beta, gamma) to positions left-to-right. That
results in a different allocation compared to the old ``sum_case``.
So, now, sumSyntax’s sum_case_tm has alpha as the range type, and beta
as the domain of the first function, and gamma as the domain of the
second, as before.

The code for sumSyntax.mk_sum_case should probably just use
list_mk_icomb for the sake of robustness.