Fix bug in parsing of case-expressions involving “literals”.
Issue arises when the parsing code thinks that it’s seeing a literal,
which it does on seeing a “if var = exp then e1 else e2” under a case
expression, even when exp includes free variables. (The same buggy
behaviour was seen before the case-const arg-flipping work began, but
it had to be a bool_case term then.)

The fix is to make the parsing code check that the exp in expressions
such as the above really does have no free variables. Then it can
treat that as a literal case expression.