Fix for bug identified in dc26673.
Problem was that TypeBase’s congruence rule for the case constant over
booleans was not the same as COND_CONG, and the former got applied
rather than the latter when a function’s r.h.s. was being rewritten.

Previously, this was not an issue because the case constant over
booleans wasn’t the same term as COND. But when these two got
identified, TFL started to use

   > Prim_rec.case_cong_thm BOOL_CASES_AX bool_case_thm;
   val it =
      |- ∀M M' t1 t2.
        (M ⇔ M') ∧ ((M' ⇔ T) ⇒ (t1 = t1')) ∧ ((M' ⇔ F) ⇒ (t2 = t2')) ⇒
        ((if M then t1 else t2) = if M' then t1' else t2')

as the congruence rule, putting unnecessary = T, and = F stuff into
induction principles for functions.