Allow avoid-token-merging code to print strings asserted to be zero-width.
It’s conceivable that a client may want to output a string to the
pretty-printing stream that the client wishes the stream to view as
having no width. The code previously didn’t allow this. Now, the code
checks to see if the desired output is actually the empty string
before doing nothing.