Make antiquotes in grammar definition quotations stand for token sets.
As per the example at the bottom of grammarLib.sml, this allows for things like

  `T ::= ^(``{Ident s | s ≠ "" ∧ isAlpha (HD s)}``)
      |  "(" E ")";`

which allows for tokens to be a subset of some broader class. If you
wanted the whole set, you could write

  `T ::= <Ident> | "(" E ")";`

This is done entirely to support the miniML example where special
tokens with infix status like "*" are still lexed as SymbolT "*", not
as a separate StarT token, say. This then means that normal variables
cannot be "*", and to reflect this you need a rule like

  `V ::= ^(``{SymbolT s | s ≠ "*" ∧ s ≠ "/" ...}``);`