This changes the rank of a case constant for a newly defined datatype.
It affects the possible rank of the bodies that can be used when
building case statements with that case constant. The bodies may
be of any rank up to or equalling the rank of the case constant.
Previously, the bodies could only have rank 1.

This is still a temporary solution; ideally the bodies should be
able to be of any rank, but apparently this would require an infinity
of case constants of increasing ranks, which is too ugly.