Partial fix for issue #127.
The definition of "remove_parse_stack" now works for the "classic heuristic" (set using PmatchHeuristics) but still fails for the "default". For the time being, an argument could be making the "classic" the default again, since it appears to be more stable.

The pertinent change was in Defn.wfrec_eqns, where "elim_triv_literal_CONV" is now incorporated into "rule". Other changes are tidy-ups.