fix possible bug in Vimhol mode
The hol.vim file was already designed to be able to handle scripts that
look like:
  \\ tac1 \\ tac2
  \\ tac3
i.e. with the THEN combinator appearing at the start of the line rather
than the end. However, the call to vim's search function to find and
strip off the leading '\\' was somehow stripping off too much.

The \ze in the search pattern indicating the end of the match was
perhaps only attached to the branch matching stripBothWords. When that
branch wasn't matched, the end of the match became the end of the outer
\%( \), i.e. after the delim. (And in my tests vim wanted to go even
further, to the end of the line.) In the new search pattern, the \ze
comes just before the delim and after the stuff to strip, no matter
which branch was matched.