Make Poly/ML Holmake output something when emulating HOLMOSMLC etc.
When a makefile rule wasn't a special one featuring $(HOLMOSMLC), the
command was output (just as in regular make), but when the emulation
kicked in, the "compilation" happened silently.

The messages are not a copy of the mosml command-line (which is not
being run), but rather a reflection of the arguments being passed to
the underlying code.