Update to patriciaLib code.
Following these updates, most conversions are now faster.

** Note that PTREE_TRANSFORM_CONV, PTREE_EVERY_LEAF_CONV and PTREE_EXISTS_LEAF_CONV now take a conversion as an extra argument.

** The reference is_ptree_term_size_limit has been removed.

Instead, there are now two functions: Define_mk_ptree and Define_mk_ptree_with_is_ptree. The latter returns an IS_PTREE theorem for the new constant. In this case the IS_PTREE theorem is proven in a more efficient manner, however it can still be time consuming, e.g. on a fast machine it takes 3 minutes to prove for a tree with around 49 thousand elements. Smaller trees are more manageable, e.g. 1.6s for a tree with a thousand elements.