Make listTheory's (short-lived) SINGL constant an overload to ``\x. [x]``
It's nice to have SINGL to hand as an abbreviation for the above
abstraction (particularly as writing the abstraction will usually
require enclosing parentheses, lifting the character count to 9,
compared to SINGL's 5).  But one doesn't want to see SINGL 3, when [3]
is so much nicer.

Use the new feature "ignore-for-printing" feature of overload_on (see
ac4806d), so this happens.  Somewhat sadly, this does mean that
although one can write MAP SINGL l, one will get back MAP (\x. [x]) l.