version of computeLib.write_datatype_info for other compsets
write_datatype_info is specialised for the_compset; this commit
introduces a more general version that can be applied to any compset.

Additionally, the behaviour is slightly changed: it additionally sets a
skip of SOME 1 for the datatype's case constant (if any). Nobody knows
what set_skip does, but this appears not to have broken the build.