add rulename option to LaTeX munger for naming inference rules
Allows the user to provide a string (name) to be passed as the optional
argument to \infer[name]{...} when typesetting using rule or
stackedrule.

Due to the simplistic parsing of the munger's commands and options,
unexpected results are likely if this name contains any special
characters. (Perhaps it should parse them under some encoding for this
option?) The name is wrapped in a new \HOLRuleName command, which
provides a route for customisation.