Handle corrupt Systeml.sig when moving build from Moscow ML to Poly.
This is analogous to what was done in b151b8f0 (Jul 2009), which coped
with building a Moscow ML setup on top of a build that used to be
Poly. There the claim was made that the reverse situation (the one
fixed in this commit) didn't need anything done to it. That was
because poly's Systeml.ui was being generated afresh with every
configure, right up until 331d321e (Jan 2014).