Clean up some doubly-nested LaTeX environments for the sake of Emacs.
The problem is not the double-nesting per se, but the use of

    \begin{foo}\begin{bar}
    ...
    \end{bar}\end{foo}

which, when bar is "alltt" or "verbatim" seems to cause AucTeX to get
into a terrible font-locking confusion.

Also remove some unnecessary local definitions of \bk to be a teletype
backslash when our commands.tex include-file has \bs for this
character already.