Clean up pred_set/Manual a little.
Prompted by ptroja's commit in cd6153e7.

Note that this material is desperately in need of a more comprehensive
update. The changes here are no more than a hint of how we might more
forward with this (there are similar remnants elsewhere).

In particular, the theorems in pred_set/help/thms should be generated,
and should probably be using EmitTeX technology rather than verbatim
rendering. See also github issue #41. The make process does not
correctly incorporate the manual style .doc entries, and those files
are probably wrong to boot.