Substitute implicit_rewrites for basic_rewrites in docs for Rewrite
Thanks to Piotr Trojanek for finding the problem: basic_rewrites is
now called implicit_rewrites (and is a function to boot; maybe it was
a magic value in HOL88). There isn't any documentation of
implicit_rewrites, but its nature as a function from unit is implicit
in some of the examples in the docfiles.

Closes #195