More accurate dependencies in Holmakefiles for Moscow ML.
In particular, with the change in 6e6ca26, some Holmakefiles feature
selftest.exe as the first thing to be built. If the dependencies for
these files are not recorded accurately, then the build will fail when
Moscow ML attempts to link object files to create the executable.

In src/TeX the problem was that a .uo file didn't record its
dependency on the corresponding .ui file, and again, the change in
6e6ca26 meant that targets were attempted in a different order.

The problem doesn't arise with Poly/ML because there is no analogue of
linking, and calls to the "compiler" don't actually do anything much.