do not export constant t0 from wotTheory
I get the impression that none of the definitions in wotTheory were
supposed to be exported, so perhaps they all should be deleted at the
end. I am not sure, though, so I just remove t0, whose name is
particularly annoying outside this theory to be taken by a constant.