Get Poly/Holmake to apply the quotation filter to more files.
The Moscow ML implementation just applies it to everything (and
perhaps the Poly/ML implementation should do the same). The new
behaviour in the Poly/ML code is to apply the filter if the `
character is detected, or if any character with code > 127 appears in
the file to be consumed.

Curiously, the quotation filter is applied within Holmake under Moscow
ML, but within the load implementation in Poly/ML, and this lives in
tools-poly/poly/poly-init2.ML. Because it isn't used in
tools-poly/Holmake.sml, I can delete some redundant code in

The test-case exercises this because it uses only "new style"
quotations (enabled by 199f6f07d0).