Start to rationalise Holdep's interface.
There were some strange holdovers still in there from the ancient days
when it was a separate command-line tool. Aim is to ultimately drop
Lexer.lex and Parser.grm and replace those with the Holdep_tokens
lexer. I'm also only doing this to the Moscow ML implementation for
the moment with the ultimate aim being to have the Poly and mosml
implementations using the same Holdep code.