Allow Define to work even if = has been overloaded
Doing so is probably a bad idea, but if the quotation would parse
successfully to a term, it should work as a definition too. Issue is
that the use of pdest_eq required a preterm of Const("="), but when
"=" is overloaded the preterm will be using the Overloaded constructor
instead.

This form of definition used to work, incidentally.  It was broken by
work on issue #112