Modify term-encoding in theory files to be more compact.
Both kernels now use a reverse-polish encoding (with the two binary
operators being application and abstraction). There is no whitespace
or any parentheses in the new encoding. On "typical" core theories
such as integerTheory and wordsTheory, this saves about 10% in total
theory size. On larger theories (such as those in CakeML), I hope the
saving will be even greater.